Fix bug.

+++ modules/character_generation/deblur_keyframe.py
@@ -17,7 +17,7 @@ import logging
 from typing import List
 import random
 
-from common.constant import TASK_DIR, VIDEO_DIR, MANUAL_DIR
+from common.constant import TASK_DIR, VIDEO_DIR, MANUAL_DIR, CHARACTER_DIR
 from modules.character_generation.repair_hands import deblur_image
 
 logging.basicConfig(level=logging.INFO, format="%(asctime)s [%(filename)s:%(lineno)s][%(levelname)s] %(message)s")
@@ -29,12 +29,14 @@ def deblur_keyframe(keyframe_path: Path) -> None:
     video_id, character_id = task_dir.name, task_dir.parent.name
     output_dir = MANUAL_DIR / "keyframe_deblur_selection" / f"{character_id}_{video_id}"
     output_dir.mkdir(parents=True, exist_ok=True)
+    prompt_file = CHARACTER_DIR / character_id / "prompt.json"
+    with open(prompt_file, "r") as f:
+        prompt = json.load(f).get("deblur_prompt", "")
     for i in range(20):
         seed = random.randint(0, 1000000)
-        deblur_image(keyframe_path, output_dir / f"deblur_{i}_{seed}.png", seed=seed)
+        deblur_image(keyframe_path, output_dir / f"deblur_{i}_{seed}.png", prompt=prompt, seed=seed)
 
 def extract_keyframe(task_dir: str, character_id: str, video_id: str, skip_if_exists: bool = True) -> None:
-    tasks = []
     if task_dir:
         assert character_id == "" and video_id == "", "task_dir is not allowed to be used with character_id or video_id"
         task_dir = Path(task_dir)
@@ -42,16 +44,10 @@ def extract_keyframe(task_dir: str, character_id: str, video_id: str, skip_if_ex
         character_id = task_dir.parent.name
     candidate_characters = sorted(TASK_DIR.glob("*")) \
         if not character_id else [TASK_DIR / character_id]
-    collected_keyframes = []
     for character_dir in candidate_characters:
         candidate_videos = sorted(character_dir.glob("*")) \
             if not video_id else [character_dir / video_id]
         for video_dir in candidate_videos:
-            output_dir = video_dir / "4_keyframe"
-            if output_dir.exists() and skip_if_exists:
-                logging.info(f"Key frame already processed: {output_dir}")
-                continue
-            output_dir.mkdir(parents=True)
             video_id = video_dir.name
             video_info_path = VIDEO_DIR / video_id / "video_info.json"
             if not video_info_path.exists():
@@ -69,6 +65,11 @@ def extract_keyframe(task_dir: str, character_id: str, video_id: str, skip_if_ex
                 logger.info(f"Key frame image not found: {video_dir}")
                 continue
 
+            output_dir = video_dir / "4_keyframe"
+            if output_dir.exists() and skip_if_exists:
+                logging.info(f"Key frame already processed: {output_dir}")
+                continue
+            output_dir.mkdir(parents=True, exist_ok=True)
             logger.info(f"Deblurring keyframe: {key_frame_image}")
             shutil.copy(key_frame_image, output_dir / "keyframe_origin.png")
             deblur_keyframe(output_dir / "keyframe_origin.png")
@@ -78,6 +79,6 @@ if __name__ == "__main__":
     parser.add_argument("--task_dir", type=str, required=False)
     parser.add_argument("--character_id", type=str, required=False)
     parser.add_argument("--video_id", type=str, required=False)
-    parser.add_argument("--skip_if_exists", type=bool, default=True)
+    parser.add_argument("--skip_if_exists", type=bool, default=False)
     args = parser.parse_args()
     extract_keyframe(args.task_dir, args.character_id, args.video_id, args.skip_if_exists)
